dom/tests/mochitest/general/test_idleapi_permissions.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for idle api permissions</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

var idleObserver = {
  onidle: null,
  onactive: null
};

function run_test() {
  // addIdleObserver checks whether time is > 0.
  this.idleObserver.time = 100;

  var added = false;
  try {
    navigator.addIdleObserver(this.idleObserver);
    added = true;
  } catch (e) { }
  ok(!added, "Should not be able to add idle observer without permission");
  
  SpecialPowers.addPermission("idle", true, document);
  added = false;
  try {
    navigator.addIdleObserver(this.idleObserver);
    added = true;
  } catch (e) { }
  ok(added, "Should be able to add idle observer with permission.");
  
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(run_test);
</script>
</pre>
</body>
</html>
